홈 메인 페이지 진행중 HomeController 중복되서 내일 같이 수정해야해~~

+++ java/vo/JobopenVO.java
@@ -14,13 +14,13 @@ public class JobopenVO {
 	private String jobopen_cimg;
 	private MultipartFile jobopen_cimgf;
 	private String jobopen_count;
-	private int endDday;
+	private long endDday;
 	
 	
-	public int getEndDday() {
+	public long getEndDday() {
 		return endDday;
 	}
-	public void setEndDday(int endDday) {
+	public void setEndDday(long endDday) {
 		this.endDday = endDday;
 	}
 	public int getJobopen_id() {
